timetag_parse: Use skip_wraps=1 by default
Not sure why this wasn't the default previously. Regardless, this
eliminates the large timestamp offset that timetag files were seeing.

In principle this shouldn't have affected anything, but unfortunately it
seems this offset upsets favia, resulting in large G(\infnty) values for
reasons I haven't had a chance to investigate.