anisotropy: Work-around for convolution of long models
For decay models with long lifetimes there were previously artifacts due
to the model not converging to zero by the end of the interval over
which we sampled it (which was previously the excitation period). This
caused spikey behavior in the convolved model near the end of the
excitation period.

We work around this by decoupling the model sampling interval from the
excitation period (currently ten times the excitation period), computing
the convolution over a far longer time series than should be necessary,
and dropping everything but the samples corresponding to the observed
decay samples.