favia: Fix computation of offset time
We have known for a while now that favia exhibits a number of odd
behaviors when faced with timestamps which start far from 0. This is a
bug and should be fixed but in the meantime we have been working around
this by substracting the time offset from the timestamps before handing
them to favia.

Unfortunately, the previous correction had the potential to be very
wrong as it did this,

    x -= x[0]
    y -= y[0]

In the case of low count-rates there is potentially a large gap between
x[0] and y[0]. In subtracting these offsets we have essentially offset
the time axes of the timeseries relative to one another. This
results in quite bizarre artifacts in the produced correlation function.